requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.0
python-dotenv==1.0.1
requests==2.32.3
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
